add trainInfo get and pagenation

+++ dotnet_ts_support/Models/TrainInfoModel.cs
@@ -14,10 +14,10 @@ namespace dotnet_ts_support.Models
         public string status { get; init; }
         public int progress { get; init; }
         public string createdAt { get; init; }
-        public int train_loss { get; init; }
-        public int test_loss { get; init; }
-        public int test_accuracy { get; init; }
-        public int iou { get; init; }
+        public double train_loss { get; init; }
+        public double test_loss { get; init; }
+        public double test_accuracy { get; init; }
+        public double iou { get; init; }
         public int iteration { get; init; }
         public int max_iteration { get; init; }
         public string directoryId { get; init; }
